This commit introduces massive concurrency in this emulator.
The graphics and sound systems have been moved into a separate thread. It was
running at 60hz, but that's currently disabled due to bugs.  Now it just runs as
fast as it can.

Sound is running in a separate thread as before.

There is a separate thread for timer countdown as before.

There is now a thread_sync struct to signal shutdown to all the threads so
proper shutdown and cleanup occurs.

The interactive debug UI interactions with system simulation have been revamped
to accomodate the new multithreaded design.

This is still in-progress as there are responsiveness issues and presentation issues.